function that returns the number of bits you would need to flip to get from one number to another

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,7 +11,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned int count;
-	unsigned int diff;
+	unsigned long int diff;
 
 	count = 0;
 	diff = n ^ m;
